@@ -2,47 +2,56 @@
   "name": "recoverysigner-firebase-auth",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": ">=20.0.0"
+  },
   "dependencies": {
-    "@babel/core": "^7.10.5",
-    "@babel/helper-call-delegate": "^7.11.4",
-    "@babel/helper-validator-option": "^7.14.5",
-    "@lingui/babel-preset-react": "^2.9.1",
-    "@lingui/cli": "^2.9.1",
-    "@lingui/macro": "^2.9.1",
-    "@lingui/react": "^2.9.1",
-    "@sentry/browser": "^6.16.1",
-    "@stellar/eslint-config": "^1.0.3",
-    "@stellar/prettier-config": "^1.0.1",
-    "@stellar/tsconfig": "^1.0.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.0.4",
-    "@testing-library/user-event": "^12.1.6",
-    "@types/jest": "^26.0.4",
-    "@types/lingui__macro": "^2.7.3",
-    "@types/lingui__react": "^2.8.2",
-    "@types/lodash": "^4.14.150",
-    "@types/node": "^14.11.2",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-redux": "^7.1.7",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-plugin-macros": "^2.8.0",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-react": "^1.1.7",
-    "firebase": "^9.6.2",
+    "@sentry/browser": "^8.51.0",
+    "firebase": "^11.2.0",
     "lodash": "^4.17.21",
-    "prettier": "^2.0.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-scripts": "3.4.3",
-    "reduce-reducers": "^1.0.4",
-    "typescript": "4.0.3"
+    "react": "19",
+    "react-dom": "19",
+    "react-redux": "^9.2.0",
+    "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "firebase-tools": "^10.0.1",
-    "husky": "^9.1.7"
+    "@babel/core": "^7.26.7",
+    "@babel/helper-call-delegate": "^7.12.13",
+    "@babel/helper-validator-option": "^7.25.9",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@lingui/babel-preset-react": "^2.9.1",
+    "@lingui/cli": "^5.1.2",
+    "@lingui/macro": "^5.1.2",
+    "@lingui/react": "^5.1.2",
+    "@stellar/prettier-config": "^1.2.0",
+    "@stellar/tsconfig": "^1.0.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^29.5.14",
+    "@types/lingui__macro": "^3.0.0",
+    "@types/lingui__react": "^3.0.0",
+    "@types/lodash": "^4.17.14",
+    "@types/node": "^22.10.10",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@types/react-redux": "^7.1.34",
+    "@typescript-eslint/eslint-plugin": "^8.21.0",
+    "@typescript-eslint/parser": "^8.21.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-plugin-macros": "^3.1.0",
+    "eslint": "^9.19.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-config-react": "^1.1.7",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "firebase-tools": "^13.29.2",
+    "husky": "^9.1.7",
+    "prettier": "^3.4.2",
+    "react-scripts": "5.0.1",
+    "reduce-reducers": "^1.0.4",
+    "typescript": "5.7.3"
   },
   "scripts": {
     "start": "react-scripts start",
