@@ -1,5 +1,4 @@
-import { Catalog, setupI18n } from "@lingui/core";
-import { t } from "@lingui/macro";
+import { setupI18n } from "@lingui/core";
 
 import en from "locales/en/messages";
 import es from "locales/es/messages";
@@ -9,7 +8,8 @@ import pt from "locales/pt/messages";
 import uk from "locales/uk/messages";
 import ru from "locales/ru/messages";
 
-export const catalogs: { [key: string]: Catalog } = {
+// Combine all catalogs into a single object
+const catalogs = {
   en,
   es,
   "es-AR": esAR,
@@ -19,12 +19,22 @@ export const catalogs: { [key: string]: Catalog } = {
   ru,
 };
 
-export const i18n = setupI18n({
-  language: "en",
-  catalogs,
+// Create the i18n instance
+const i18n = setupI18n();
+
+// Load all language catalogs into the i18n instance
+Object.entries(catalogs).forEach(([locale, messages]) => {
+  i18n.load(locale, messages);
 });
 
+// Set the default language
+i18n.activate("en");
+
 // Strings that need to be translated because they're being used on index.html
-t`Please update Android WebView to start the session.`;
-t`Please update Google Chrome to start the session.`;
-t`Your device’s browser is not compatible. Please update your browser and Vibrant to the latest version.`;
+i18n._("Please update Android WebView to start the session.");
+i18n._("Please update Google Chrome to start the session.");
+i18n._(
+  "Your device's browser is not compatible. Please update your browser and Vibrant to the latest version.",
+);
+
+export { catalogs, i18n };
